Update filesrc/main/java/com/example/demo/service/TaskService.java
Add field private final TaskLogsRepository repositoryLogs
Update conctructor
Update method deleteTask

+++ src/main/java/com/example/demo/service/TaskService.java
@@ -2,6 +2,7 @@ package com.example.demo.service;
 
 import com.example.demo.model.Task;
 import com.example.demo.repository.TaskRepository;
+import com.example.demo.repository.TaskLogsRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
@@ -11,9 +12,11 @@ import java.util.Optional;
 @Service
 public class TaskService {
     private final TaskRepository repository;
+    private final TaskLogsRepository repositoryLogs;
 
-    public TaskService(TaskRepository repository) {
+    public TaskService(TaskRepository repository, TaskLogsRepository repositoryLogs) {
         this.repository = repository;
+        this.repositoryLogs = repositoryLogs;
     }
 
     public List<Task> getAllTasks() {
@@ -47,6 +50,11 @@ public class TaskService {
         if (!repository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found");
         }
+        if (!repositoryLogs.existsByTaskId(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "TaskLogs: task_id not found");
+        }
+        Long taskLogId = repositoryLogs.getIdByTaskId(id);
+        repositoryLogs.deleteById(taskLogId);
         repository.deleteById(id);
     }
 }
